move utils into database module
